backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
sqlalchemy==2.0.35
psycopg2-binary==2.9.9
python-dotenv==1.0.1
bcrypt==4.2.0
PyJWT==2.9.0
pydantic[email]==2.9.2
